game works but when u click play again the game doesn't restart 2

+++ app/src/main/java/com/example/voresklimaplan/game/domain/gameTargets/GameTarget.kt
@@ -17,5 +17,5 @@ class FallingGameTarget(
 
     //kordinater
     var xCordinate: Int,
-    var yCordinate: Int
+    var yCordinate: Float
 ): GameTarget(targetName, goodForClimate, imageId,)
